third_party/blink/web_tests/platform/mac-mac15-arm64/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/reduce_l2.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] [required] reduceL2 float32 1D constant tensor empty axes
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceL2 float32 actual -4.860228061676025 should be close enough to expected 4.860228061676025 by ULP distance: expected a number less than or equal to 4n but got 2167868922n
Harness: the test ran to completion.
